refactor methods into abstract class so they can be used correctly
